static시작
